db: Change in memory blkdb database to use data from LMDB databases instead of from file Also rename blkdb database to chaindb to avoid confusion with LMDB blockdb
