+++ include/ccoin/Makefile.am
@@ -39,7 +39,7 @@ ccoininclude_HEADERS = \
 ccoindbincludedir=$(includedir)/ccoin/db
 
 ccoindbinclude_HEADERS = \
-    db/blkdb.h  \
+    db/chaindb.h  \
     db/db.h
 
 ccoinnetincludedir=$(includedir)/ccoin/net
