+++ test/chaindb.c
@@ -2,13 +2,12 @@
  * Distributed under the MIT/X11 software license, see the accompanying
  * file COPYING or http://www.opensource.org/licenses/mit-license.php.
  */
-#include "picocoin-config.h"
 
 #include <ccoin/buffer.h>               // for const_buffer
 #include <ccoin/buint.h>                // for hex_bu256, bu256_copy, etc
 #include <ccoin/core.h>                 // for bp_block_calc_sha256, etc
 #include <ccoin/coredefs.h>             // for chain_info, chain_metadata, etc
-#include <ccoin/db/blkdb.h>             // for blkinfo, blkdb, blkdb_reorg, etc
+#include <ccoin/db/chaindb.h>           // for blkinfo, chaindb, chaindb_reorg, etc
 #include <ccoin/db/db.h>                // for blockdb_init, etc
 #include <ccoin/key.h>                  // for bp_key_static_shutdown
 #include <ccoin/log.h>                  // for logging
@@ -23,7 +22,7 @@
 
 struct logging *log_state;
 
-static void add_header(struct blkdb *db, char *raw)
+static void add_header(struct chaindb *db, char *raw)
 {
 	struct const_buffer buf = { raw, 80 };
 
@@ -36,15 +35,15 @@ static void add_header(struct blkdb *db, char *raw)
 
 	bu256_copy(&bi->hash, &bi->hdr.sha256);
 
-	struct blkdb_reorg reorg;
+	struct chaindb_reorg reorg;
 
-	assert(blkdb_add(db, bi, &reorg) == true);
+	assert(chaindb_add(db, bi, &reorg) == true);
 
 	assert(reorg.conn == 1);
 	assert(reorg.disconn == 0);
 }
 
-static void read_headers(const char *ser_base_fn, struct blkdb *db)
+static void read_headers(const char *ser_base_fn, struct chaindb *db)
 {
 	char *filename = test_filename(ser_base_fn);
 	int fd = file_seq_open(filename);
@@ -60,7 +59,7 @@ static void read_headers(const char *ser_base_fn, struct blkdb *db)
 	free(filename);
 }
 
-static void test_blkinfo_prev(struct blkdb *db)
+static void test_blkinfo_prev(struct chaindb *db)
 {
 	struct blkinfo *tmp = db->best_chain;
 	int height = db->best_chain->height;
@@ -78,13 +77,13 @@ static void test_blkinfo_prev(struct blkdb *db)
 static void runtest(const char *ser_base_fn, const struct chain_info *chain,
 		    unsigned int check_height, const char *check_hash)
 {
-	struct blkdb db;
+	struct chaindb db;
 
 	bu256_t block0;
 	bool rc = hex_bu256(&block0, chain->genesis_hash);
 	assert(rc);
 
-	rc = blkdb_init(&db, chain->netmagic, &block0);
+	rc = chaindb_init(&db, chain->netmagic, &block0);
 	assert(rc);
 
 	read_headers(ser_base_fn, &db);
@@ -98,7 +97,7 @@ static void runtest(const char *ser_base_fn, const struct chain_info *chain,
 
 	test_blkinfo_prev(&db);
 
-	blkdb_free(&db);
+	chaindb_free(&db);
 }
 
 int main (int argc, char *argv[])
